EdgeBackGestureHandler: use tick instead of heavy tick for back gesture
Now that the duplicated AOSP behavior is removed in commit 1e698690,
the haptic feedback toggle for back gesture finally determines
whether there is a haptic feedback or not.

Since the toggle description says "haptic tick effect" not "haptic
heavy tick effect", change the actual behavior to use EFFECT_TICK
as well. EFFECT_TICK is also the default behavior from AOSP, and
users are used to the light tick from previous builds. Let's not
suprise users with a sudden heavy tick that gets triggered with
back gestures hundreds of times a day.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: Cykeek <Cykeek@proton.me>